backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
pydantic==2.9.2
pydantic-settings==2.6.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
openpyxl==3.1.5
pandas==2.2.3
xlrd==2.0.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
